Calculate the PLC50 for a given P~abs~ and PAH chemical using the PTLM

PLC50 is the LC50 of a phototoxic PAH based on calculations of site-specific or field-level light absorption.

Usage

plc50(x, pah = NULL, NLC50 = NULL)

Arguments

x

light absorption, calculated from p_abs(), or a tuv_results data.frame from tuv() or get_tuv_results().

pah

The PAH of interest, which is used to look up the NLC50.

NLC50

(optional) the narcotic toxicity (i.e., in the absence of light) of the PAH in ug/L. If supplied, takes precedence over the PAH lookup.

Value

the PLC50 of the PAH in ug/L.

Details

You can either supply a specific PAH, so the NLC50 can be calculated for that chemical, or supply a NLC50 value directly.

References

Marzooghi, S., Finch, B.E., Stubblefield, W.A., Dmitrenko, O., Neal, S.L. and Di Toro, D.M. (2017), Phototoxic target lipid model of single polycyclic aromatic hydrocarbons. Environ Toxicol Chem, 36: 926-937. https://doi.org/10.1002/etc.3601

Examples

plc50(590, pah = "Benzo[a]pyrene")
#> [1] 0.05979498
plc50(590, NLC50 = 450)
#> [1] 14.68394